+++ packages/webaudio/index.mjs
@@ -6,3 +6,5 @@ This program is free software: you can redistribute it and/or modify it under th
 
 export * from './webaudio.mjs';
 export * from './sampler.mjs';
+export * from './helpers.mjs';
+export * from './synth.mjs';
