+++ packages/webaudio/sampler.mjs
@@ -1,5 +1,6 @@
 import { logger, toMidi, valueToMidi } from '@strudel.cycles/core';
-import { getAudioContext } from './index.mjs';
+import { getAudioContext, setSound } from './index.mjs';
+import { getADSR } from './helpers.mjs';
 
 const bufferCache = {}; // string: Promise<ArrayBuffer>
 const loadCache = {}; // string: Promise<ArrayBuffer>
@@ -20,7 +21,7 @@ function humanFileSize(bytes, si) {
   return bytes.toFixed(1) + ' ' + units[u];
 }
 
-export const getSampleBufferSource = async (s, n, note, speed, freq) => {
+export const getSampleBufferSource = async (s, n, note, speed, freq, bank) => {
   let transpose = 0;
   if (freq !== undefined && note !== undefined) {
     logger('[sampler] hap has note and freq. ignoring note', 'warning');
@@ -29,23 +30,6 @@ export const getSampleBufferSource = async (s, n, note, speed, freq) => {
   transpose = midi - 36; // C3 is middle C
 
   const ac = getAudioContext();
-  // is sample from loaded samples(..)
-  const samples = getLoadedSamples();
-  if (!samples) {
-    throw new Error('no samples loaded');
-  }
-  const bank = samples?.[s];
-  if (!bank) {
-    throw new Error(
-      `sample not found: "${s}"`,
-      // , try one of ${Object.keys(samples)
-      // .map((s) => `"${s}"`)
-      // .join(', ')}.
-    );
-  }
-  if (typeof bank !== 'object') {
-    throw new Error('wrong format for sample bank:', s);
-  }
   let sampleUrl;
   if (Array.isArray(bank)) {
     sampleUrl = bank[n % bank.length];
@@ -107,8 +91,6 @@ export const getLoadedBuffer = (url) => {
   return bufferCache[url];
 };
 
-let sampleCache = { current: undefined };
-
 /**
  * Loads a collection of samples to use with `s`
  * @example
@@ -147,37 +129,129 @@ export const samples = async (sampleMap, baseUrl = sampleMap._base || '') => {
         throw new Error(`error loading "${sampleMap}"`);
       });
   }
-  sampleCache.current = {
-    ...sampleCache.current,
-    ...Object.fromEntries(
-      Object.entries(sampleMap).map(([key, value]) => {
-        if (typeof value === 'string') {
-          value = [value];
-        }
-        if (typeof value !== 'object') {
-          throw new Error('wrong sample map format for ' + key);
-        }
-        baseUrl = value._base || baseUrl;
-        const replaceUrl = (v) => (baseUrl + v).replace('github:', 'https://raw.githubusercontent.com/');
-        if (Array.isArray(value)) {
-          return [key, value.map(replaceUrl)];
-        }
-        // must be object
-        return [
-          key,
-          Object.fromEntries(
-            Object.entries(value).map(([note, samples]) => {
-              return [note, (typeof samples === 'string' ? [samples] : samples).map(replaceUrl)];
-            }),
-          ),
-        ];
-      }),
-    ),
-  };
+  Object.entries(sampleMap).forEach(([key, value]) => {
+    if (typeof value === 'string') {
+      value = [value];
+    }
+    if (typeof value !== 'object') {
+      throw new Error('wrong sample map format for ' + key);
+    }
+    baseUrl = value._base || baseUrl;
+    const replaceUrl = (v) => (baseUrl + v).replace('github:', 'https://raw.githubusercontent.com/');
+    if (Array.isArray(value)) {
+      //return [key, value.map(replaceUrl)];
+      value = value.map(replaceUrl);
+    } else {
+      // must be object
+      value = Object.fromEntries(
+        Object.entries(value).map(([note, samples]) => {
+          return [note, (typeof samples === 'string' ? [samples] : samples).map(replaceUrl)];
+        }),
+      );
+    }
+    setSound(key, (options) => onTriggerSample(options, value));
+  });
 };
 
-export const resetLoadedSamples = () => {
-  sampleCache.current = undefined;
-};
+const cutGroups = [];
 
-export const getLoadedSamples = () => sampleCache.current;
+export async function onTriggerSample(options, bank) {
+  const { hap, duration: hapDuration, t, cps } = options;
+  const {
+    s,
+    freq,
+    unit,
+    nudge = 0, // TODO: is this in seconds?
+    cut,
+    loop,
+    clip = 0, // if 1, samples will be cut off when the hap ends
+    n = 0,
+    note,
+    speed = 1, // sample playback speed
+    begin = 0,
+    end = 1,
+  } = hap.value;
+  const ac = getAudioContext();
+  // destructure adsr here, because the default should be different for synths and samples
+  const { attack = 0.001, decay = 0.001, sustain = 1, release = 0.001 } = hap.value;
+  // load sample
+  if (speed === 0) {
+    // no playback
+    return;
+  }
+  if (!s) {
+    // is this check really needed?
+    console.warn('no sample specified');
+    return;
+  }
+  //const soundfont = getSoundfontKey(s);
+  let bufferSource;
+
+  //if (soundfont) {
+  // is soundfont
+  //bufferSource = await globalThis.getFontBufferSource(soundfont, note || n, ac, freq);
+  //} else {
+  // is sample from loaded samples(..)
+  bufferSource = await getSampleBufferSource(s, n, note, speed, freq, bank);
+  //}
+  // asny stuff above took too long?
+  if (ac.currentTime > t) {
+    logger(`[sampler] still loading sound "${s}:${n}"`, 'highlight');
+    // console.warn('sample still loading:', s, n);
+    return;
+  }
+  if (!bufferSource) {
+    console.warn('no buffer source');
+    return;
+  }
+  bufferSource.playbackRate.value = Math.abs(speed) * bufferSource.playbackRate.value;
+  if (unit === 'c') {
+    // are there other units?
+    bufferSource.playbackRate.value = bufferSource.playbackRate.value * bufferSource.buffer.duration * cps;
+  }
+  const shouldClip = /* soundfont || */ clip;
+  let duration = shouldClip ? hapDuration : bufferSource.buffer.duration / bufferSource.playbackRate.value;
+  // "The computation of the offset into the sound is performed using the sound buffer's natural sample rate,
+  // rather than the current playback rate, so even if the sound is playing at twice its normal speed,
+  // the midway point through a 10-second audio buffer is still 5."
+  const offset = begin * duration * bufferSource.playbackRate.value;
+  duration = (end - begin) * duration;
+  if (loop) {
+    bufferSource.loop = true;
+    bufferSource.loopStart = offset;
+    bufferSource.loopEnd = offset + duration;
+    duration = loop * duration;
+  }
+  const time = t + nudge;
+
+  bufferSource.start(time, offset);
+  if (cut !== undefined) {
+    cutGroups[cut]?.stop(time); // fade out?
+    cutGroups[cut] = bufferSource;
+  }
+  //chain.push(bufferSource);
+  bufferSource.stop(t + duration + release);
+  const adsr = getADSR(attack, decay, sustain, release, 1, time, time + duration);
+  bufferSource.connect(adsr);
+  //chain.push(adsr);
+  return adsr;
+}
+
+/*const getSoundfontKey = (s) => {
+  if (!globalThis.soundfontList) {
+    // soundfont package not loaded
+    return false;
+  }
+  if (globalThis.soundfontList?.instruments?.includes(s)) {
+    return s;
+  }
+  // check if s is one of the soundfonts, which are loaded into globalThis, to avoid coupling both packages
+  const nameIndex = globalThis.soundfontList?.instrumentNames?.indexOf(s);
+  // convert number nameIndex (0-128) to 3 digit string (001-128)
+  const name = nameIndex < 10 ? `00${nameIndex}` : nameIndex < 100 ? `0${nameIndex}` : nameIndex;
+  if (nameIndex !== -1) {
+    // TODO: indices of instrumentNames do not seem to match instruments
+    return globalThis.soundfontList.instruments.find((instrument) => instrument.startsWith(name));
+  }
+  return;
+};*/
