+++ packages/webaudio/webaudio.mjs
@@ -5,15 +5,21 @@ This program is free software: you can redistribute it and/or modify it under th
 */
 
 import * as strudel from '@strudel.cycles/core';
-import { fromMidi, logger, toMidi } from '@strudel.cycles/core';
 import './feedbackdelay.mjs';
 import './reverb.mjs';
-import { getSampleBufferSource } from './sampler.mjs';
 const { Pattern } = strudel;
 import './vowel.mjs';
 import workletsUrl from './worklets.mjs?url';
+import { getFilter, gainNode } from './helpers.mjs';
 
 // export const getAudioContext = () => Tone.getContext().rawContext;
+
+export const soundMap = new Map();
+// onTrigger = ({ hap: Hap, t: number, deadline: number, duration: number, cps: number }) => AudioNode
+export function setSound(key, onTrigger) {
+  soundMap.set(key, onTrigger);
+}
+export const resetLoadedSounds = () => soundMap.clear();
 
 let audioContext;
 export const getAudioContext = () => {
@@ -36,65 +42,6 @@ const getDestination = () => {
 export const panic = () => {
   getDestination().gain.linearRampToValueAtTime(0, getAudioContext().currentTime + 0.01);
   destination = null;
-};
-
-const getFilter = (type, frequency, Q) => {
-  const filter = getAudioContext().createBiquadFilter();
-  filter.type = type;
-  filter.frequency.value = frequency;
-  filter.Q.value = Q;
-  return filter;
-};
-
-const getADSR = (attack, decay, sustain, release, velocity, begin, end) => {
-  const gainNode = getAudioContext().createGain();
-  gainNode.gain.setValueAtTime(0, begin);
-  gainNode.gain.linearRampToValueAtTime(velocity, begin + attack); // attack
-  gainNode.gain.linearRampToValueAtTime(sustain * velocity, begin + attack + decay); // sustain start
-  gainNode.gain.setValueAtTime(sustain * velocity, end); // sustain end
-  gainNode.gain.linearRampToValueAtTime(0, end + release); // release
-  // for some reason, using exponential ramping creates little cracklings
-  /* let t = begin;
-  gainNode.gain.setValueAtTime(0, t);
-  gainNode.gain.exponentialRampToValueAtTime(velocity, (t += attack));
-  const sustainGain = Math.max(sustain * velocity, 0.001);
-  gainNode.gain.exponentialRampToValueAtTime(sustainGain, (t += decay));
-  if (end - begin < attack + decay) {
-    gainNode.gain.cancelAndHoldAtTime(end);
-  } else {
-    gainNode.gain.setValueAtTime(sustainGain, end);
-  }
-  gainNode.gain.exponentialRampToValueAtTime(0.001, end + release); // release */
-  return gainNode;
-};
-
-const getOscillator = ({ s, freq, t, duration, release }) => {
-  // make oscillator
-  const o = getAudioContext().createOscillator();
-  o.type = s || 'triangle';
-  o.frequency.value = Number(freq);
-  o.start(t);
-  o.stop(t + duration + release);
-  return o;
-};
-
-const getSoundfontKey = (s) => {
-  if (!globalThis.soundfontList) {
-    // soundfont package not loaded
-    return false;
-  }
-  if (globalThis.soundfontList?.instruments?.includes(s)) {
-    return s;
-  }
-  // check if s is one of the soundfonts, which are loaded into globalThis, to avoid coupling both packages
-  const nameIndex = globalThis.soundfontList?.instrumentNames?.indexOf(s);
-  // convert number nameIndex (0-128) to 3 digit string (001-128)
-  const name = nameIndex < 10 ? `00${nameIndex}` : nameIndex < 100 ? `0${nameIndex}` : nameIndex;
-  if (nameIndex !== -1) {
-    // TODO: indices of instrumentNames do not seem to match instruments
-    return globalThis.soundfontList.instruments.find((instrument) => instrument.startsWith(name));
-  }
-  return;
 };
 
 let workletsLoading;
@@ -135,13 +82,6 @@ export async function initAudioOnFirstClick() {
     });
   });
 }
-
-function gainNode(value) {
-  const node = getAudioContext().createGain();
-  node.gain.value = value;
-  return node;
-}
-const cutGroups = [];
 
 let delays = {};
 function getDelay(orbit, delaytime, delayfeedback, t) {
@@ -188,13 +128,9 @@ export const webaudioOutput = async (hap, deadline, hapDuration, cps) => {
   let t = ac.currentTime + deadline;
   // destructure value
   let {
-    freq,
-    s,
+    s = 'triangle',
     bank,
-    sf,
-    clip = 0, // if 1, samples will be cut off when the hap ends
-    n = 0,
-    note,
+    source,
     gain = 0.8,
     // low pass
     cutoff,
@@ -210,17 +146,10 @@ export const webaudioOutput = async (hap, deadline, hapDuration, cps) => {
     crush,
     shape,
     pan,
-    speed = 1, // sample playback speed
-    begin = 0,
-    end = 1,
     vowel,
     delay = 0,
     delayfeedback = 0.5,
     delaytime = 0.25,
-    unit,
-    nudge = 0, // TODO: is this in seconds?
-    cut,
-    loop,
     orbit = 1,
     room,
     size = 2,
@@ -232,87 +161,13 @@ export const webaudioOutput = async (hap, deadline, hapDuration, cps) => {
   if (bank && s) {
     s = `${bank}_${s}`;
   }
-  if (!s || ['sine', 'square', 'triangle', 'sawtooth'].includes(s)) {
-    // destructure adsr here, because the default should be different for synths and samples
-    const { attack = 0.001, decay = 0.05, sustain = 0.6, release = 0.01 } = hap.value;
-    // with synths, n and note are the same thing
-    n = note || n || 36;
-    if (typeof n === 'string') {
-      n = toMidi(n); // e.g. c3 => 48
-    }
-    // get frequency
-    if (!freq && typeof n === 'number') {
-      freq = fromMidi(n); // + 48);
-    }
-    // make oscillator
-    const o = getOscillator({ t, s, freq, duration: hapDuration, release });
-    chain.push(o);
-    // level down oscillators as they are really loud compared to samples i've tested
-    chain.push(gainNode(0.3));
-    // TODO: make adsr work with samples without pops
-    // envelope
-    const adsr = getADSR(attack, decay, sustain, release, 1, t, t + hapDuration);
-    chain.push(adsr);
+  if (soundMap.has(s)) {
+    const node = await soundMap.get(s)({ hap, t, deadline, duration: hapDuration, cps });
+    chain.push(node);
+  } else if (source) {
+    chain.push(source({ hap, t, deadline, duration: hapDuration, cps }));
   } else {
-    // destructure adsr here, because the default should be different for synths and samples
-    const { attack = 0.001, decay = 0.001, sustain = 1, release = 0.001 } = hap.value;
-    // load sample
-    if (speed === 0) {
-      // no playback
-      return;
-    }
-    if (!s) {
-      console.warn('no sample specified');
-      return;
-    }
-    const soundfont = getSoundfontKey(s);
-    let bufferSource;
-
-    if (soundfont) {
-      // is soundfont
-      bufferSource = await globalThis.getFontBufferSource(soundfont, note || n, ac, freq);
-    } else {
-      // is sample from loaded samples(..)
-      bufferSource = await getSampleBufferSource(s, n, note, speed, freq);
-    }
-    // asny stuff above took too long?
-    if (ac.currentTime > t) {
-      logger(`[sampler] still loading sound "${s}:${n}"`, 'highlight');
-      // console.warn('sample still loading:', s, n);
-      return;
-    }
-    if (!bufferSource) {
-      console.warn('no buffer source');
-      return;
-    }
-    bufferSource.playbackRate.value = Math.abs(speed) * bufferSource.playbackRate.value;
-    if (unit === 'c') {
-      // are there other units?
-      bufferSource.playbackRate.value = bufferSource.playbackRate.value * bufferSource.buffer.duration * cps;
-    }
-    let duration = soundfont || clip ? hapDuration : bufferSource.buffer.duration / bufferSource.playbackRate.value;
-    // "The computation of the offset into the sound is performed using the sound buffer's natural sample rate,
-    // rather than the current playback rate, so even if the sound is playing at twice its normal speed,
-    // the midway point through a 10-second audio buffer is still 5."
-    const offset = begin * duration * bufferSource.playbackRate.value;
-    duration = (end - begin) * duration;
-    if (loop) {
-      bufferSource.loop = true;
-      bufferSource.loopStart = offset;
-      bufferSource.loopEnd = offset + duration;
-      duration = loop * duration;
-    }
-    t += nudge;
-
-    bufferSource.start(t, offset);
-    if (cut !== undefined) {
-      cutGroups[cut]?.stop(t); // fade out?
-      cutGroups[cut] = bufferSource;
-    }
-    chain.push(bufferSource);
-    bufferSource.stop(t + duration + release);
-    const adsr = getADSR(attack, decay, sustain, release, 1, t, t + duration);
-    chain.push(adsr);
+    throw new Error(`sound ${s} not found! Is it loaded?`);
   }
 
   // gain stage
